+++ index.js
@@ -1,1 +1,1 @@
-module.exports = require('./lib');
+module.exports = require('./src');
