+++ src/helpers/convertGlobPaths.js
