+++ src/helpers/finalizeSpecificationObject.js
