+++ src/helpers/getSpecificationObject.js
