+++ src/helpers/hasEmptyProperty.js
