+++ src/helpers/parseApiFile.js
