+++ src/helpers/parseApiFileContent.js
