+++ src/helpers/updateSpecificationObject.js
